refactor: 이메일 및 닉네임 중복확인 리팩토링

+++ src/main/java/com/dart/api/application/member/MemberService.java
@@ -1,9 +1,14 @@
 package com.dart.api.application.member;
 
+import static java.lang.Boolean.*;
+
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
+
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.RequiredArgsConstructor;
 
 import com.dart.api.domain.auth.entity.AuthUser;
 import com.dart.api.domain.member.entity.Member;
@@ -12,12 +17,14 @@ import com.dart.api.dto.member.request.MemberUpdateDto;
 import com.dart.api.dto.member.request.NicknameDuplicationCheckDto;
 import com.dart.api.dto.member.request.SignUpDto;
 import com.dart.api.dto.member.response.MemberProfileResDto;
+import com.dart.api.domain.auth.repository.EmailRedisRepository;
+import com.dart.api.domain.auth.repository.NicknameRedisRepository;
+import com.dart.api.domain.auth.repository.SessionRedisRepository;
 import com.dart.api.infrastructure.s3.S3Service;
 import com.dart.global.error.exception.BadRequestException;
+import com.dart.global.error.exception.ConflictException;
 import com.dart.global.error.exception.NotFoundException;
 import com.dart.global.error.model.ErrorCode;
-
-import lombok.RequiredArgsConstructor;
 
 @Service
 @RequiredArgsConstructor
@@ -25,23 +32,33 @@ import lombok.RequiredArgsConstructor;
 public class MemberService {
 
 	private final MemberRepository memberRepository;
+	private final EmailRedisRepository emailRedisRepository;
+	private final NicknameRedisRepository nicknameRedisRepository;
+	private final SessionRedisRepository sessionRedisRepository;
+
 	private final S3Service s3Service;
-	private final NicknameValidator nicknameValidator;
+	private final NicknameService nicknameService;
 	private final PasswordEncoder passwordEncoder;
 
 	@Transactional
-	public void signUp(SignUpDto signUpDto) {
-		validateEmailChecked(signUpDto.isCheckedEmail());
-		validateNicknameChecked(signUpDto.isCheckedNickname());
+	public void signUp(SignUpDto signUpDto, String sessionId) {
+		nicknameService.setNicknameVerified(signUpDto.nickname());
+
+		verifyEmailChecked(signUpDto.email());
+		verifyNicknameChecked(signUpDto.nickname());
+		validateMemberExistsByEmail(signUpDto.email());
 
 		final String encodedPassword = passwordEncoder.encode(signUpDto.password());
 		final Member member = Member.signup(signUpDto, encodedPassword);
 
 		memberRepository.save(member);
+
+		cleanUpSessionData(sessionId, signUpDto.email(), signUpDto.nickname());
 	}
 
 	public MemberProfileResDto getMemberProfile(String nickname, AuthUser authUser) {
-		if (isMember(authUser) && isOwnProfile(nickname, authUser.nickname())) {
+		validateMemberExistsByNickname(nickname);
+		if(isOwnProfile(authUser.nickname(), nickname)) {
 			return getOwnProfile(nickname);
 		} else {
 			return getOtherProfile(nickname);
@@ -49,29 +66,60 @@ public class MemberService {
 	}
 
 	@Transactional
-	public void updateMemberProfile(AuthUser authUser, MemberUpdateDto memberUpdateDto, MultipartFile profileImage) {
-		validateNicknameChecked(memberUpdateDto.isCheckedNickname());
+	public void updateMemberProfile(
+		AuthUser authUser,
+		MemberUpdateDto memberUpdateDto,
+		MultipartFile profileImage,
+		String sessionId) {
 
-		final Member member = findMemberByEmail(authUser.email());
-		final String savedProfileImage = member.getProfileImageUrl();
+		Member member = findMemberByEmail(authUser.email());
+		validateNickname(memberUpdateDto.nickname(), member.getNickname(), sessionId);
+		String newProfileImageUrl = handleProfileImageUpdate(profileImage, member.getProfileImageUrl());
+
+		member.updateMemberProfile(memberUpdateDto, newProfileImageUrl);
+		handleNicknameUpdate(memberUpdateDto.nickname(), member.getNickname(), sessionId);
+	}
+
+	public void checkNicknameDuplication(NicknameDuplicationCheckDto nicknameDuplicationCheckDto, String sessionId,
+		HttpServletResponse response) {
+		nicknameService.checkAndReserveNickname(nicknameDuplicationCheckDto.nickname(), sessionId, response);
+	}
+
+	private void validateNickname(String newNickname, String currentNickname, String sessionId) {
+		if (newNickname != null && !newNickname.equals(currentNickname)) {
+			if (sessionId == null || !nicknameRedisRepository.isReserved(newNickname)) {
+				throw new BadRequestException(ErrorCode.FAIL_NOT_VERIFIED_NICKNAME);
+			}
+		}
+	}
+
+	private String handleProfileImageUpdate(MultipartFile profileImage, String savedProfileImage) {
+		if (profileImage == null || profileImage.isEmpty()) return savedProfileImage;
 
 		String newProfileImageUrl = s3Service.uploadFile(profileImage);
-		if (savedProfileImage != null)
+		if (savedProfileImage != null) {
 			s3Service.deleteFile(savedProfileImage);
-		member.updateMemberProfile(memberUpdateDto, newProfileImageUrl);
+		}
 
+		return newProfileImageUrl;
 	}
 
-	public void checkNicknameDuplication(NicknameDuplicationCheckDto nicknameDuplicationCheckDto) {
-		nicknameValidator.validate(nicknameDuplicationCheckDto.nickname());
+	private void handleNicknameUpdate(String newNickname, String currentNickname, String sessionId) {
+		if (newNickname != null && !newNickname.equals(currentNickname)) {
+			sessionRedisRepository.deleteSessionNicknameMapping(sessionId);
+			nicknameRedisRepository.deleteNickname(newNickname);
+		}
 	}
 
-	private boolean isMember(AuthUser authUser) {
-		return authUser != null;
+	private void cleanUpSessionData(String sessionId, String email, String nickname) {
+		sessionRedisRepository.deleteSessionEmailMapping(sessionId);
+		sessionRedisRepository.deleteSessionNicknameMapping(sessionId);
+		emailRedisRepository.deleteEmail(email);
+		nicknameRedisRepository.deleteNickname(nickname);
 	}
 
-	private boolean isOwnProfile(String currentNickname, String profileNickname) {
-		return currentNickname.equals(profileNickname);
+	private boolean isOwnProfile(String authUserNickname, String profileNickname) {
+		return authUserNickname.equals(profileNickname);
 	}
 
 	private MemberProfileResDto getOwnProfile(String nickname) {
@@ -96,15 +144,27 @@ public class MemberService {
 			.orElseThrow(() -> new NotFoundException(ErrorCode.FAIL_MEMBER_NOT_FOUND));
 	}
 
-	private void validateEmailChecked(boolean isEmailChecked) {
-		if (!isEmailChecked) {
+	private void verifyEmailChecked(String email) {
+		if(FALSE.equals(emailRedisRepository.isVerified(email))) {
 			throw new BadRequestException(ErrorCode.FAIL_NOT_VERIFIED_EMAIL);
 		}
 	}
 
-	private void validateNicknameChecked(boolean isNicknameChecked) {
-		if (!isNicknameChecked) {
+	private void verifyNicknameChecked(String nickname) {
+		if(FALSE.equals(nicknameRedisRepository.isReserved(nickname))) {
 			throw new BadRequestException(ErrorCode.FAIL_NOT_VERIFIED_NICKNAME);
+		}
+	}
+
+	private void validateMemberExistsByEmail(String email) {
+		if (memberRepository.existsByEmail(email)) {
+			throw new ConflictException(ErrorCode.FAIL_EMAIL_CONFLICT);
+		}
+	}
+
+	private void validateMemberExistsByNickname(String nickname) {
+		if (memberRepository.existsByEmail(nickname)) {
+			throw new ConflictException(ErrorCode.FAIL_EMAIL_CONFLICT);
 		}
 	}
 }
